Get XperfUI to build for UNICODE as well as ANSI - default is now UNICODE.
This UNICODE change is a bit of a hack, since in most cases I am just forcing
ANSI functions to be used even though _UNICODE is defined. Eventually the project
should probably be changed to be truly _UNICODE or TCHAR compatible, and that
won't be a big deal, but that isn't the way I went.

Building for UNICODE is important for MFC apps because non-UNICODE builds are
deprecated by Microsoft starting with VS 2013, and static builds at least
require downloading a separate installer.