Display output of StripChromeSymbols.py while it is running.
Previously the ChildProcess class would only display output after the
process had terminated. Now the parent process can sit in a loop waiting
for process termination or fresh output, and print the output as it appears.
This keeps the output printing on the main thread but avoids the problem
of having no output for 30+ seconds at a time.