Tweak StripChromeSymbols.py and fix so that its output shows up in real-time.
When Python has its output directed to a file or pipe then it buffers, which
defeats my recent change. You need to pass -u to defeat this. Done. I then
made the print-pending-output loop part of the normal destructor, so that's
cool.

I also tweaked the overly verbose output of StripChromeSymbols.py.