Remove xperftemptracedir, fix etwrecord.bat in the morning, explicitly select "NT Kernel Logger"
I've decided that xperftemptracedir is a dumb idea. Temporary traces should be
recorded to %temp%, and users should make sure that is on a fast drive. This is
fixed in XperfUI and in etwrecord.bat, and removed it from the settings dialog.

Apparently I haven't run etwrecord.bat in the morning before because it fails before
10 am. Fixed. As a side-effect of this fix the seconds now show up in the file-name
correctly. Testing whether it also fails before 1:00 am is left as an exercise for the reader.

I also changed the default xperftracedir in etwrecord.bat to documents\xperftraces.