Internal list of traces is now stored without a file extension.
In order to simplify eventual support for zipped traces the trace names
are now stored without file extensions. This makes it easier to ignore
files that have .etl and .zip variants.
This doesn't change the visible behavior, except in the case where
there is a .zip file but not an .etl file -- XperfUI now detects and
warns about that case.