Implement trace renaming, which also renames associated files.
Simplify the GetEditControlText function.
Tweak the definition of GetFileExt() to point at the period if present,
and tweak GetFilePart() to return the entire string if no slash is present.
Add menu items for Zip compress trace and traces (not implemented.
Add menu item for rename trace.
Add an edit control that appears when trace renaming starts and disappears
later.