Look for and register Chromium repos and record Chrome ETW events.
chrome.dll and chrome_events.man are looked for in two common
repo locations. If found then chrome.dll is registered. If this
succeeds then +Chrome is added to the list of user providers.

ChildProcess was tweaked so that GetExitCode waits for process
exit - this closes off a buggy pattern that leads to race
conditions and simplifies the code. Also don't call outputPrintf
with an empty string.

Some double-spaces in comments were fixed to avoid confusing them
with spaces-as-tabs.