Various bits of cleanup, especially to ChildProcess and input logging.
The ChildProcess class was cleaned up to initialize variables more
consistently in the class definition, better document usage, and make
more functions private.

The input logging code (in KeyLoggerThread.*) was cleaned up so that
it stops the thread as needed, including at shutdown. This means that
g_LoggingEnabled isn't needed anymore. g_LogKeyboardDetails is now
an std::atomic<bool> instead of a volatile bool.

A few other removals of extraneous MFC goop.