More code cleanup - moving directory monitor to separate class.
The directory monitor thread is now a separate class in a
separate source file, in order to get the main XperfUIDlg.cpp
file as small and simple as possible.

The directory monitor thread now gets shut down cleanly when
the CXperfUIDlg object is destroyed.

A few other functions were moved inline, renamed, or deleted.