Add loading and saving of settings.
All settings on the main window are now saved and restored. The code
for this is in support.cpp, to avoid clutter in the main XperfUIDlg.cpp
file.

It seems odd that I had to write wrapper functions for Get/WriteProfileInt,
but if I hadn't then I would be specifying key names twice and variable
names three times -- madness.