Tweaks to StripChromeSymbols.py output and a hack tweak.
StripChromeSymbols.py's output was verbose but not always clear.
I changed the output to print the major commansd being executed,
using a "> " prefix to indicate that they are commands, instead of
the word-wrap inducing "Running: ".

The logic for working around amcache.hve bugs was also tweaked to
make it more resilient in the cases where UIforETW crashes or
exits while recording a trace. This really only happens to me, but
when it does it was causing ~6 minute hangs and required manual
fixup. Not fun.