Adding IdentifyChromeProcesses.py and auto run it.
IdentifyChromeProcesses.py is automatically run after a trace is
recorded. It records the Chrome PIDs and the process types in the
parallel text file, which is helpful for Chrome developers. The
information can be deleted if it is unwanted. Eventually non-Chrome
developers will be able to turn off this feature.