Allow renaming of traces that don't start with the standard date/time.
UIforETW creates traces with names that start with the date and time.
This allows for consistent sorting by date, and when renaming the date
is excluded from renaming, for simplicity. However traces from other
sources may not fit this format. In this case the renaming functionality
should allow editing the entire name. The code just looks for a long
enough string of digits and special characters. More sophisticated
checks could be done, but would add little value.