Clean up StripChromeSymbols.py to find support files reliably.
StripChromeSymbols.py relies on a number of executables and DLLs.
Some of these are stored in ..\third_party and must be copied to
the bin directory on first run. This is required because they have
to be in the same directory as the executable using them (which is
RetrieveSymbols.exe).

StripChromeSymbols.py then needs to reliably find these executables
so it now explicitly runs them from the script directory.